Moves ClientAlternateTyping to _DV
